requirements.txt
numpy
pandas
matplotlib
streamlit
python-dotenv
virtualenv
pandas
openai
azure-monitor-opentelemetry